refactor: HandlerAdapterRegistry 클래스로 adapter 찾는 로직 분리
