+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -2,8 +2,8 @@ package com.techcourse;
 
 import jakarta.servlet.ServletContext;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.ManualMappingHandlerAdapter;
-import nextstep.mvc.RequestMappingHandlerAdapter;
+import nextstep.mvc.adapter.ManualMappingHandlerAdapter;
+import nextstep.mvc.adapter.RequestMappingHandlerAdapter;
 import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
 import org.slf4j.Logger;
